% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first.R
\name{add}
\alias{add}
\title{Addition}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number to \strong{add}}

\item{y}{Another number to add}
}
\value{
x+y
}
\description{
Addition
}
\examples{
add(3,4)
add(1:3,3:7)
}
