% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter}
\alias{filter}
\title{Filter rows of a data frame}
\usage{
filter(x, condition)
}
\arguments{
\item{x}{: Data frame}

\item{condition}{: Expressions that return a logical value}
}
\value{
Subset dataframe by rows
}
\description{
Filter rows of a data frame
}
