% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec.R
\name{select}
\alias{select}
\title{Subset column of a data fram}
\usage{
select(x, var)
}
\arguments{
\item{x}{:Dataframe}

\item{var}{: Either a vector of column indices or variable names or logicals}
}
\value{
A subset data frame
}
\description{
Subset column of a data fram
}
\examples{
select(iris, 1:2)
select(iris, "Species")
select(iris, c(TRUE,FALSE,FALSE,TRUE,FALSE))

}
